webnn: add support limits for `elu` `gelu` `leakyRelu` and `relu`
This CL adds data type support limits for `elu`, `gelu`, `leakyRelu` and
`relu`, also unifies `where_true_value` and `where_false_value` into
`where_value` in mojo since they should have the same data type.

Bug: 345271830
Cq-Include-Trybots: luci.chromium.try:win11-blink-rel,mac14.arm64-blink-rel,mac14-blink-rel, linux-blink-rel
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5706672
Commit-Queue: Phillis Tang <phillis@chromium.org>
Auto-Submit: Shiyi Zou <shiyi.zou@intel.com>
Cr-Commit-Position: refs/heads/main@{#1332411}

+++ services/webnn/webnn_graph_impl_unittest.cc
@@ -53,11 +53,14 @@ ContextProperties GetContextPropertiesForTesting() {
                          /*arg_min_max_output=*/
                          {OperandDataType::kInt32, OperandDataType::kInt64},
                          /*concat_inputs=*/SupportedDataTypes::All(),
+                         /*elu_input=*/SupportedDataTypes::All(),
                          /*gather_input=*/SupportedDataTypes::All(),
                          /*gather_indices=*/SupportedDataTypes::All(),
+                         /*gelu_input=*/SupportedDataTypes::All(),
+                         /*leaky_relu_input=*/SupportedDataTypes::All(),
+                         /*relu_input=*/SupportedDataTypes::All(),
                          /*where_condition=*/SupportedDataTypes::All(),
-                         /*where_true_value=*/SupportedDataTypes::All(),
-                         /*where_false_value=*/SupportedDataTypes::All()}));
+                         /*where_value=*/SupportedDataTypes::All()}));
 }
 
 // A fake WebNNGraph Mojo interface implementation that binds a pipe for
@@ -2134,8 +2137,8 @@ TEST_F(WebNNGraphImplTest, EluTest) {
   }
   {
     // Test the invalid graph when the alpha is NAN.
-    EluTester{.input = {.type = OperandDataType::kInt32, .dimensions = {2}},
-              .output = {.type = OperandDataType::kInt32, .dimensions = {2}},
+    EluTester{.input = {.type = OperandDataType::kFloat32, .dimensions = {2}},
+              .output = {.type = OperandDataType::kFloat32, .dimensions = {2}},
               .alpha = NAN,
               .expected = false}
         .Test();
@@ -5162,10 +5165,11 @@ TEST_F(WebNNGraphImplTest, ReluTest) {
   }
   {
     // Test relu operator for 4-D tensor with int32 input.
-    ReluTester{
-        .input = {.type = OperandDataType::kInt32, .dimensions = {1, 5, 3, 7}},
-        .output = {.type = OperandDataType::kInt32, .dimensions = {1, 5, 3, 7}},
-        .expected = true}
+    ReluTester{.input = {.type = OperandDataType::kFloat32,
+                         .dimensions = {1, 5, 3, 7}},
+               .output = {.type = OperandDataType::kFloat32,
+                          .dimensions = {1, 5, 3, 7}},
+               .expected = true}
         .Test();
   }
   {
